+++ manage-server/src/main/java/manage/control/DatabaseController.java
@@ -20,6 +20,7 @@ import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
 import javax.sql.DataSource;
@@ -32,6 +33,7 @@ import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 
 @Component
+@RestController
 public class DatabaseController {
 
     private RestTemplate restTemplate;
